refactor: InsufficientPointBalanceException 클래스  이동
