+++ user/src/main/java/com/food/user/error/InsufficientPointBalanceException.java
@@ -1,7 +1,6 @@
-package com.food.user;
+package com.food.user.error;
 
 import com.food.common.error.ApplicationErrors;
-import com.food.user.error.PointErrors;
 
 public class InsufficientPointBalanceException extends RuntimeException {
     private static final ApplicationErrors ERROR = PointErrors.INSUFFICIENT_POINT_BALANCE;
