+++ user/src/test/java/com/food/user/PointUseTest.java
@@ -4,6 +4,7 @@ import com.food.common.user.business.internal.dto.PointDto;
 import com.food.common.user.business.internal.dto.UserDto;
 import com.food.common.utils.Amount;
 import com.food.common.utils.UsedPoints;
+import com.food.user.error.InsufficientPointBalanceException;
 import com.food.user.error.NotFoundPointOwnerException;
 import com.food.user.mock.MockPoint;
 import com.food.user.mock.MockRequestUser;
